requirements.txt
langchain==0.0.230
datasets[s3]==2.13.0
discord 
aiohttp
python-dotenv
eland 
elasticsearch 
huggingface-hub 
tqdm 
requests 
ipython